translate syscalls to strings

+++ libxcache/src/peek.h
@@ -16,8 +16,8 @@ static inline long peek_reg(pid_t pid, size_t offset) {
   offsetof(struct user, regs) + offsetof(struct user_regs_struct, reg)
 
 /// read the syscall number a child is in the midst of calling
-static inline long peek_syscall_no(pid_t pid) {
-  return peek_reg(pid, REG(orig_rax));
+static inline unsigned long peek_syscall_no(pid_t pid) {
+  return (unsigned long)peek_reg(pid, REG(orig_rax));
 }
 
 /** read a NUL-terminated string out of a child’s address space
